First cut MVC implementation
* Basic functionalities for making annotation is fully implemented
* Removed all deprecated classes from 0.x version
