--- src/main/java/edu/brandeis/cs/nlp/mae/controller/ColorRect.java
@@ -1,57 +1,0 @@
-/*
- * MAE - Multi-purpose Annotation Environment
- *
- * Copyright Keigh Rim (krim@brandeis.edu)
- * Department of Computer Science, Brandeis University
- * Original program by Amber Stubbs (astubbs@cs.brandeis.edu)
- *
- * MAE is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, @see <a href="http://www.gnu.org/licenses">http://www.gnu.org/licenses</a>.
- *
- * For feedback, reporting bugs, use the project repo on github
- * @see <a href="https://github.com/keighrim/mae-annotation">https://github.com/keighrim/mae-annotation</a>
- */
-
-package edu.brandeis.cs.nlp.mae.controller;
-
-import javax.swing.*;
-import java.awt.*;
-
-/**
- * filled color icon class to be used in tab titles as toggle buttons
- */
-public class ColorRect implements Icon {
-    private int size;
-    private Color color;
-
-    public ColorRect(Color c, int size) {
-        this.color = c;
-        this.size = size;
-    }
-
-    @Override
-    public void paintIcon(Component c, Graphics g, int x, int y) {
-        g.setColor(this.color);
-        g.fillRect(x, y, this.size, this.size);
-    }
-
-    @Override
-    public int getIconWidth() {
-        return this.size;
-    }
-
-    @Override
-    public int getIconHeight() {
-        return this.size;
-    }
-}
