--- src/main/java/edu/brandeis/cs/nlp/mae/controller/deprecated/FontSizeMenuListener.java
@@ -1,55 +1,0 @@
-/*
- * MAE - Multi-purpose Annotation Environment
- *
- * Copyright Keigh Rim (krim@brandeis.edu)
- * Department of Computer Science, Brandeis University
- * Original program by Amber Stubbs (astubbs@cs.brandeis.edu)
- *
- * MAE is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, @see <a href="http://www.gnu.org/licenses">http://www.gnu.org/licenses</a>.
- *
- * For feedback, reporting bugs, use the project repo on github
- * @see <a href="https://github.com/keighrim/mae-annotation">https://github.com/keighrim/mae-annotation</a>
- */
-
-package edu.brandeis.cs.nlp.mae.controller.deprecated;
-
-import edu.brandeis.cs.nlp.mae.controller.MaeMainUI;
-
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-/**
- * Class that changes the size of the text from the top menu
- */
-public class FontSizeMenuListener implements ActionListener {
-    private MaeMainUI maeMainUI;
-
-    public FontSizeMenuListener(MaeMainUI maeMainUI) {
-        this.maeMainUI = maeMainUI;
-    }
-
-    public void actionPerformed(ActionEvent actionEvent) {
-        String command = actionEvent.getActionCommand();
-        if (command.equals("Font++")) {
-            Font font = maeMainUI.getTextPanel().getFont();
-            Font font2 = new Font(font.getName(), font.getStyle(), font.getSize() + 1);
-            maeMainUI.getTextPanel().setFont(font2);
-        } else if (command.equals("Font--")) {
-            Font font = maeMainUI.getTextPanel().getFont();
-            Font font2 = new Font(font.getName(), font.getStyle(), font.getSize() - 1);
-            maeMainUI.getTextPanel().setFont(font2);
-        }
-    }
-}
