--- src/main/java/edu/brandeis/cs/nlp/mae/controller/deprecated/HelpMenuListener.java
@@ -1,65 +1,0 @@
-/*
- * MAE - Multi-purpose Annotation Environment
- *
- * Copyright Keigh Rim (krim@brandeis.edu)
- * Department of Computer Science, Brandeis University
- * Original program by Amber Stubbs (astubbs@cs.brandeis.edu)
- *
- * MAE is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, @see <a href="http://www.gnu.org/licenses">http://www.gnu.org/licenses</a>.
- *
- * For feedback, reporting bugs, use the project repo on github
- * @see <a href="https://github.com/keighrim/mae-annotation">https://github.com/keighrim/mae-annotation</a>
- */
-
-package edu.brandeis.cs.nlp.mae.controller.deprecated;
-
-import edu.brandeis.cs.nlp.mae.MaeStrings;
-import edu.brandeis.cs.nlp.mae.controller.MaeMainUI;
-
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-
-/**
- * Listens for the request from the Help Menu
- */
-public class HelpMenuListener implements ActionListener {
-    private MaeMainUI maeMainUI;
-
-    public HelpMenuListener(MaeMainUI maeMainUI) {
-        this.maeMainUI = maeMainUI;
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent actionEvent) {
-        String command = actionEvent.getActionCommand();
-        if (command.equals("about")) {
-            maeMainUI.showAboutDialog();
-        } else if (command.equals("web")) {
-            if (Desktop.isDesktopSupported()) {
-                try {
-                    Desktop.getDesktop().browse(
-                            new URI(MaeStrings.PROJECT_WEBPAGE));
-                } catch (IOException e) {
-                    e.printStackTrace();
-                } catch (URISyntaxException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-}
